chang float to double

+++ ch2_snake_model/include/gradient_descent_base.h
@@ -3,7 +3,7 @@
 
 class GradientDescentBase {
    public:
-    GradientDescentBase(float step_size);
+    GradientDescentBase(double step_size);
     void run(int max_iteration);
 
    protected:
@@ -11,13 +11,13 @@ class GradientDescentBase {
     virtual void update() = 0;
     virtual bool is_terminate(int current_iter, int max_iteration) const;
 
-    virtual float compute_energy() = 0;
+    virtual double compute_energy() = 0;
     virtual void roll_back_state() = 0;
     virtual void back_up_state() = 0;
     virtual void update_step_size(bool is_energy_decent);
 
     virtual void print_terminate_info() const;
 
-    float step_size_ = 1e-10;
-    float last_energy_ = std::numeric_limits<float>::max();
+    double step_size_ = 1e-10;
+    double last_energy_ = std::numeric_limits<double>::max();
 };
